[PB-1638] refactor MoveItemDialog breadcrumbs
